requirements.txt
Flask==0.12.0
discord.py
asyncio
